Feat: Get Enum from BE

+++ web_admin_fitness/lib/modules/main/modules/programs/widgets/program_search_bar.dart
@@ -77,7 +77,7 @@ class _ProgramSearchBarState extends State<ProgramSearchBar> {
           (b) => b
             ..field = 'Program.level'
             ..operator = GFILTER_OPERATOR.Gin
-            ..data = filterData.levels.map((e) => e.toInt()).join(','),
+            ..data = filterData.levels.map((e) => e).join(','),
         ),
       );
     }
